Reverted counts to using Int64 values rather than NativeInt.

+++ RunningStatistics/Countmap.cs
@@ -11,20 +11,20 @@ namespace RunningStatistics;
 /// of zero, however a new key will not be added to the internal dictionary.
 /// </summary>
 /// <typeparam name="T">The observation type.</typeparam>
-public class Countmap<T> : IRunningStatistic<T>, IReadOnlyDictionary<T, nint>
+public class Countmap<T> : IRunningStatistic<T>, IReadOnlyDictionary<T, long>
 {
-    private readonly IDictionary<T, nint> _counter;
+    private readonly IDictionary<T, long> _counter;
 
 
     public Countmap()
     {
-        _counter = new Dictionary<T, nint>();
+        _counter = new Dictionary<T, long>();
         Count = 0;
     }
 
     public Countmap(Countmap<T> other)
     {
-        _counter = new Dictionary<T, nint>(other._counter);
+        _counter = new Dictionary<T, long>(other._counter);
         Count = other.Count;
     }
 
@@ -42,7 +42,7 @@ public class Countmap<T> : IRunningStatistic<T>, IReadOnlyDictionary<T, nint>
         Fit(value, 1);
     }
 
-    public void Fit(T obs, nint k)
+    public void Fit(T obs, long k)
     {
         Count += k;
 
@@ -81,18 +81,18 @@ public class Countmap<T> : IRunningStatistic<T>, IReadOnlyDictionary<T, nint>
 
 
 
-    public nint Count { get; private set; }
+    public long Count { get; private set; }
     public T Mode => _counter.Aggregate((x, y) => x.Value > y.Value ? x : y).Key;
-    public nint this[T key] => _counter.TryGetValue(key, out var value) ? value : 0;
+    public long this[T key] => _counter.TryGetValue(key, out var value) ? value : 0;
     public IEnumerable<T> Keys => _counter.Keys;
-    public IEnumerable<nint> Values => _counter.Values;
+    public IEnumerable<long> Values => _counter.Values;
     
     
     /// <summary>
     /// This will always return <c>true</c>. It will either return the count for the given key, or return 0.
     /// For this reason, you should use an indexer and check if the count is non-zero.
     /// </summary>
-    bool IReadOnlyDictionary<T, nint>.TryGetValue(T key, out nint value)
+    bool IReadOnlyDictionary<T, long>.TryGetValue(T key, out long value)
     {
         if (!_counter.TryGetValue(key, out value))
         {
@@ -111,15 +111,15 @@ public class Countmap<T> : IRunningStatistic<T>, IReadOnlyDictionary<T, nint>
     /// <summary>
     /// Return the counter as a <see cref="SortedDictionary{TKey,TValue}"/>, sorted by the keys.
     /// </summary>
-    public SortedDictionary<T, nint> AsSortedDictionary() => new(_counter);
+    public SortedDictionary<T, long> AsSortedDictionary() => new(_counter);
 
     /// <summary>
     /// Please use the <see cref="Count"/> property as it uses an Int64 value to store the count. This method will
     /// throw an exception if the count is too big for an Int32.
     /// </summary>
-    int IReadOnlyCollection<KeyValuePair<T, nint>>.Count => Convert.ToInt32(Count);
+    int IReadOnlyCollection<KeyValuePair<T, long>>.Count => Convert.ToInt32(Count);
     
-    public IEnumerator<KeyValuePair<T, nint>> GetEnumerator() => _counter.GetEnumerator();
+    public IEnumerator<KeyValuePair<T, long>> GetEnumerator() => _counter.GetEnumerator();
     IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
     
     public override string ToString() => $"{typeof(Countmap<T>)}(n={Count})";
